Change learning rate to float parameter

+++ src/main.py
@@ -33,7 +33,7 @@ def parse_arg():
                         help="Use Tensorflow for the optimization step")
     parser.add_argument('--optimizer', '-o', type=str, help='Optimizer type')
 
-    parser.add_argument('--lr', type=int, default=1,
+    parser.add_argument('--lr', type=float, default=1,
                         help='Learning rate interval in log10')
     parser.add_argument('--steps', type=int, default=2000,
                         help='Number of steps')
